2025.9

+++ common/src/main/java/io/github/maritims/advent_of_code/common/geometry/Grid2D.java
@@ -61,7 +61,7 @@ public class Grid2D<T> {
         return cols;
     }
 
-    public T get(long row, long col) {
+    public T get(double row, double col) {
         return grid.get((int) row).get((int) col);
     }
